Modified how makeTranscriptDb() generates internal exon ids to accomodate Ensembl annotations.
When the 'exonId' arg is supplied, now makeTranscriptDb() uses it to generate
the internal exon ids instead of trying to generate them based on the exon
loci. This is compatible with the way Ensembl assigns exon ids (in their
annotations, 2 exons with the same genomic locus can be assigned 2 distinct
ids).


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicFeatures@43907 bc3139a8-67e5-0310-9ffc-ced21a209358